feature: StartVerbatimBox and EndVerbatimBox
Added the StartVerbatimBox and EndVerbatimBox instructions.
The code between the start and end instructions will not be modified by
the compiler, thus delineating a "verbatim" box of code.

The astute reader will notice that this is merely a renaming of the
instructions StartVerbatimBlock and EndVerbatimBlock; this change is
marked as a feature to emphasize the incompatibility with v1.3.0.